requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
future==0.18.2
gkeepapi==0.11.8
gpsoauth==0.4.1
idna==2.8
pycryptodomex==3.9.4
python-dateutil==2.8.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
